fix: persist errors (wip)

+++ service/persist/postgres/user.go
@@ -28,7 +28,7 @@ type UserRepository struct {
 	getByIDsStmt             *sql.Stmt
 	getByWalletIDStmt        *sql.Stmt
 	getByUsernameStmt        *sql.Stmt
-	getByEmailStmt           *sql.Stmt
+	getByVerifiedEmailStmt   *sql.Stmt
 	deleteStmt               *sql.Stmt
 	getGalleriesStmt         *sql.Stmt
 	updateCollectionsStmt    *sql.Stmt
@@ -39,7 +39,6 @@ type UserRepository struct {
 	deleteWalletStmt         *sql.Stmt
 	addFollowerStmt          *sql.Stmt
 	removeFollowerStmt       *sql.Stmt
-	followsUserStmt          *sql.Stmt
 }
 
 // NewUserRepository creates a new postgres repository for interacting with users
@@ -51,19 +50,24 @@ func NewUserRepository(db *sql.DB, queries *db.Queries, pgx *pgxpool.Pool) *User
 	updateInfoStmt, err := db.PrepareContext(ctx, `UPDATE users SET USERNAME = $2, USERNAME_IDEMPOTENT = $3, LAST_UPDATED = $4, BIO = $5 WHERE ID = $1;`)
 	checkNoErr(err)
 
-	getByIDStmt, err := db.PrepareContext(ctx, `SELECT ID,DELETED,VERSION,USERNAME,USERNAME_IDEMPOTENT,BIO,TRAITS,WALLETS,UNIVERSAL,PRIMARY_WALLET_ID,CREATED_AT,LAST_UPDATED,EMAIL_VERIFIED FROM users WHERE ID = $1 AND DELETED = false;`)
+	// TODO update sql schema
+	getByIDStmt, err := db.PrepareContext(ctx, `SELECT ID,DELETED,VERSION,USERNAME,USERNAME_IDEMPOTENT,BIO,TRAITS,WALLETS,UNIVERSAL,PRIMARY_WALLET_ID,CREATED_AT,LAST_UPDATED FROM users WHERE ID = $1 AND DELETED = false;`)
 	checkNoErr(err)
 
-	getByIDsStmt, err := db.PrepareContext(ctx, `SELECT ID,DELETED,VERSION,USERNAME,USERNAME_IDEMPOTENT,BIO,TRAITS,WALLETS,UNIVERSAL,PRIMARY_WALLET_ID,CREATED_AT,LAST_UPDATED,EMAIL_VERIFIED FROM users WHERE ID = ANY($1) AND DELETED = false;`)
+	// TODO update sql schema
+	getByIDsStmt, err := db.PrepareContext(ctx, `SELECT ID,DELETED,VERSION,USERNAME,USERNAME_IDEMPOTENT,BIO,TRAITS,WALLETS,UNIVERSAL,PRIMARY_WALLET_ID,CREATED_AT,LAST_UPDATED FROM users WHERE ID = ANY($1) AND DELETED = false;`)
 	checkNoErr(err)
 
-	getByWalletIDStmt, err := db.PrepareContext(ctx, `SELECT ID,DELETED,VERSION,USERNAME,USERNAME_IDEMPOTENT,WALLETS,BIO,TRAITS,UNIVERSAL,PRIMARY_WALLET_ID,CREATED_AT,LAST_UPDATED,EMAIL_VERIFIED FROM users WHERE ARRAY[$1]::varchar[] <@ WALLETS AND DELETED = false;`)
+	// TODO update sql schema
+	getByWalletIDStmt, err := db.PrepareContext(ctx, `SELECT ID,DELETED,VERSION,USERNAME,USERNAME_IDEMPOTENT,WALLETS,BIO,TRAITS,UNIVERSAL,PRIMARY_WALLET_ID,CREATED_AT,LAST_UPDATED FROM users WHERE ARRAY[$1]::varchar[] <@ WALLETS AND DELETED = false;`)
 	checkNoErr(err)
 
-	getByUsernameStmt, err := db.PrepareContext(ctx, `SELECT ID,DELETED,VERSION,USERNAME,USERNAME_IDEMPOTENT,WALLETS,BIO,TRAITS,UNIVERSAL,PRIMARY_WALLET_ID,CREATED_AT,LAST_UPDATED,EMAIL_VERIFIED FROM users WHERE USERNAME_IDEMPOTENT = $1 AND DELETED = false;`)
+	// TODO update sql schema
+	getByUsernameStmt, err := db.PrepareContext(ctx, `SELECT ID,DELETED,VERSION,USERNAME,USERNAME_IDEMPOTENT,WALLETS,BIO,TRAITS,UNIVERSAL,PRIMARY_WALLET_ID,CREATED_AT,LAST_UPDATED FROM users WHERE USERNAME_IDEMPOTENT = $1 AND DELETED = false;`)
 	checkNoErr(err)
 
-	getByEmailStmt, err := db.PrepareContext(ctx, `SELECT ID,DELETED,VERSION,USERNAME,USERNAME_IDEMPOTENT,WALLETS,BIO,TRAITS,UNIVERSAL,PRIMARY_WALLET_ID,CREATED_AT,LAST_UPDATED,EMAIL_VERIFIED FROM pii.user_view WHERE PII_EMAIL_ADDRESS = $1 AND DELETED = false;`)
+	// TODO update sql schema
+	getByVerifiedEmailStmt, err := db.PrepareContext(ctx, `SELECT ID,DELETED,VERSION,USERNAME,USERNAME_IDEMPOTENT,WALLETS,BIO,TRAITS,UNIVERSAL,PRIMARY_WALLET_ID,CREATED_AT,LAST_UPDATED FROM pii.user_view WHERE PII_VERIFIED_EMAIL_ADDRESS = $1 AND DELETED = false;`)
 	checkNoErr(err)
 
 	deleteStmt, err := db.PrepareContext(ctx, `UPDATE users SET DELETED = TRUE WHERE ID = $1;`)
@@ -91,7 +95,7 @@ func NewUserRepository(db *sql.DB, queries *db.Queries, pgx *pgxpool.Pool) *User
 		getByIDsStmt:             getByIDsStmt,
 		getByWalletIDStmt:        getByWalletIDStmt,
 		getByUsernameStmt:        getByUsernameStmt,
-		getByEmailStmt:           getByEmailStmt,
+		getByVerifiedEmailStmt:   getByVerifiedEmailStmt,
 		deleteStmt:               deleteStmt,
 		getWalletIDStmt:          getWalletIDStmt,
 		getWalletStmt:            getWalletStmt,
@@ -146,6 +150,13 @@ func (u *UserRepository) createWalletWithTx(ctx context.Context, queries *db.Que
 		queries = u.queries
 	}
 
+	/*	// TODO L1Chain
+		wallet, err := queries.GetWalletByAddressAndL1Chain(ctx, db.GetWalletByAddressAndL1ChainParams{
+			Address: chainAddress.Address(),
+			L1Chain: chainAddress.Chain().L1Chain(),
+		})
+	*/
+
 	wallet, err := queries.GetWalletByChainAddress(ctx, db.GetWalletByChainAddressParams{
 		Address: chainAddress.Address(),
 		Chain:   chainAddress.Chain(),
@@ -156,14 +167,16 @@ func (u *UserRepository) createWalletWithTx(ctx context.Context, queries *db.Que
 	walletID := wallet.ID
 
 	if walletID != "" {
-
 		user, err := queries.GetUserByWalletID(ctx, walletID.String())
 		if err != nil && !errors.Is(err, pgx.ErrNoRows) {
 			return "", err
 		}
 
-		// If the wallet belongs to a user, its address can't be used to create a new wallet. Return an error.
 		if user.ID != "" {
+			if user.ID == userID {
+				// Wallet already belongs to the user; do nothing
+				return walletID, nil
+			}
 			if user.Universal {
 				err = queries.DeleteUserByID(ctx, user.ID)
 				if err != nil {
@@ -192,6 +205,8 @@ func (u *UserRepository) createWalletWithTx(ctx context.Context, queries *db.Que
 		Chain:      chainAddress.Chain(),
 		WalletType: walletType,
 		UserID:     userID,
+		// TODO L1Chain
+		// L1Chain:    chainAddress.Chain().L1Chain(),
 	})
 	if err != nil {
 		return "", persist.ErrWalletCreateFailed{
@@ -242,6 +257,19 @@ func (u *UserRepository) Create(pCtx context.Context, pUser persist.CreateUserIn
 		return "", err
 	}
 
+	/* TODO privy
+	   if pUser.PrivyDID != nil {
+	   		err := queries.SetPrivyDIDForUser(pCtx, db.SetPrivyDIDForUserParams{
+	   			ID:       persist.GenerateID(),
+	   			UserID:   userID,
+	   			PrivyDid: *pUser.PrivyDID,
+	   		})
+	   		if err != nil {
+	   			return "", err
+	   		}
+	   	}
+	*/
+
 	if pUser.ChainAddress.Address() != "" {
 		_, err = u.createWalletWithTx(pCtx, queries, pUser.ChainAddress, pUser.WalletType, userID)
 		if err != nil {
@@ -249,17 +277,29 @@ func (u *UserRepository) Create(pCtx context.Context, pUser persist.CreateUserIn
 		}
 	}
 
-	if pUser.Email != nil {
-		err := queries.UpdateUserEmail(pCtx, db.UpdateUserEmailParams{
-			UserID:                  userID,
-			EmailAddress:            *pUser.Email,
-			EmailVerificationStatus: int32(pUser.EmailStatus),
-		})
-		if err != nil {
-			logger.For(pCtx).Errorf("failed to insert email address when creating new user with userID=%s\n", userID)
-		}
-	}
+	/*
+	   TODO email
+	   if pUser.Email != nil {
+	   		if pUser.EmailStatus == persist.EmailVerificationStatusVerified {
+	   			err := queries.UpdateUserVerifiedEmail(pCtx, db.UpdateUserVerifiedEmailParams{
+	   				UserID:       userID,
+	   				EmailAddress: *pUser.Email,
+	   			})
+	   			if err != nil {
+	   				logger.For(pCtx).Errorf("failed to insert verified email address when creating new user with userID=%s\n", userID)
+	   			}
+	   		} else if pUser.EmailStatus == persist.EmailVerificationStatusUnverified {
+	   			err := queries.UpdateUserUnverifiedEmail(pCtx, db.UpdateUserUnverifiedEmailParams{
+	   				UserID:       userID,
+	   				EmailAddress: *pUser.Email,
+	   			})
+	   			if err != nil {
+	   				logger.For(pCtx).Errorf("failed to insert unverified email address when creating new user with userID=%s\n", userID)
+	   			}
+	   		}
 
+	   	}
+	*/
 	return userID, nil
 }
 
@@ -268,7 +308,7 @@ func (u *UserRepository) GetByID(pCtx context.Context, pID persist.DBID) (persis
 
 	user := persist.User{}
 	walletIDs := []persist.DBID{}
-	err := u.getByIDStmt.QueryRowContext(pCtx, pID).Scan(&user.ID, &user.Deleted, &user.Version, &user.Username, &user.UsernameIdempotent, &user.Bio, &user.Traits, pq.Array(&walletIDs), &user.Universal, &user.PrimaryWalletID, &user.CreationTime, &user.LastUpdated, &user.EmailVerified)
+	err := u.getByIDStmt.QueryRowContext(pCtx, pID).Scan(&user.ID, &user.Deleted, &user.Version, &user.Username, &user.UsernameIdempotent, &user.Bio, &user.Traits, pq.Array(&walletIDs), &user.Universal, &user.PrimaryWalletID, &user.CreationTime, &user.LastUpdated)
 	if err != nil {
 		if err == sql.ErrNoRows {
 			return persist.User{}, persist.ErrUserNotFound{UserID: pID}
@@ -304,7 +344,7 @@ func (u *UserRepository) GetByIDs(pCtx context.Context, pIDs []persist.DBID) ([]
 	for rows.Next() {
 		user := persist.User{}
 		walletIDs := []persist.DBID{}
-		rows.Scan(&user.ID, &user.Deleted, &user.Version, &user.Username, &user.UsernameIdempotent, &user.Bio, &user.Traits, pq.Array(&walletIDs), &user.Universal, &user.PrimaryWalletID, &user.CreationTime, &user.LastUpdated, &user.EmailVerified)
+		rows.Scan(&user.ID, &user.Deleted, &user.Version, &user.Username, &user.UsernameIdempotent, &user.Bio, &user.Traits, pq.Array(&walletIDs), &user.Universal, &user.PrimaryWalletID, &user.CreationTime, &user.LastUpdated)
 		wallets := make([]persist.Wallet, len(walletIDs))
 
 		for i, walletID := range walletIDs {
@@ -323,13 +363,13 @@ func (u *UserRepository) GetByIDs(pCtx context.Context, pIDs []persist.DBID) ([]
 }
 
 // GetByChainAddress gets the user who owns the wallet with the specified ChainAddress (if any)
-func (u *UserRepository) GetByChainAddress(pCtx context.Context, pChainAddress persist.ChainAddress) (persist.User, error) {
+func (u *UserRepository) GetByChainAddress(pCtx context.Context, pChainAddress persist.L1ChainAddress) (persist.User, error) {
 	var walletID persist.DBID
 
-	err := u.getWalletIDStmt.QueryRowContext(pCtx, pChainAddress.Address(), pChainAddress.Chain()).Scan(&walletID)
+	err := u.getWalletIDStmt.QueryRowContext(pCtx, pChainAddress.Address(), pChainAddress.L1Chain()).Scan(&walletID)
 	if err != nil {
 		if err == sql.ErrNoRows {
-			return persist.User{}, persist.ErrWalletNotFound{ChainAddress: pChainAddress}
+			return persist.User{}, persist.ErrWalletNotFoundByAddress{Address: pChainAddress}
 		}
 		return persist.User{}, err
 	}
@@ -341,7 +381,7 @@ func (u *UserRepository) GetByChainAddress(pCtx context.Context, pChainAddress p
 func (u *UserRepository) GetByWalletID(pCtx context.Context, pWalletID persist.DBID) (persist.User, error) {
 
 	var user persist.User
-	err := u.getByWalletIDStmt.QueryRowContext(pCtx, pWalletID).Scan(&user.ID, &user.Deleted, &user.Version, &user.Username, &user.UsernameIdempotent, pq.Array(&user.Wallets), &user.Bio, &user.Traits, &user.Universal, &user.PrimaryWalletID, &user.CreationTime, &user.LastUpdated, &user.EmailVerified)
+	err := u.getByWalletIDStmt.QueryRowContext(pCtx, pWalletID).Scan(&user.ID, &user.Deleted, &user.Version, &user.Username, &user.UsernameIdempotent, pq.Array(&user.Wallets), &user.Bio, &user.Traits, &user.Universal, &user.PrimaryWalletID, &user.CreationTime, &user.LastUpdated)
 	if err != nil {
 		if err == sql.ErrNoRows {
 			return persist.User{}, persist.ErrUserNotFound{WalletID: pWalletID}
@@ -369,7 +409,7 @@ func (u *UserRepository) GetByWalletID(pCtx context.Context, pWalletID persist.D
 func (u *UserRepository) GetByUsername(pCtx context.Context, pUsername string) (persist.User, error) {
 
 	var user persist.User
-	err := u.getByUsernameStmt.QueryRowContext(pCtx, strings.ToLower(pUsername)).Scan(&user.ID, &user.Deleted, &user.Version, &user.Username, &user.UsernameIdempotent, pq.Array(&user.Wallets), &user.Bio, &user.Traits, &user.Universal, &user.PrimaryWalletID, &user.CreationTime, &user.LastUpdated, &user.EmailVerified)
+	err := u.getByUsernameStmt.QueryRowContext(pCtx, strings.ToLower(pUsername)).Scan(&user.ID, &user.Deleted, &user.Version, &user.Username, &user.UsernameIdempotent, pq.Array(&user.Wallets), &user.Bio, &user.Traits, &user.Universal, &user.PrimaryWalletID, &user.CreationTime, &user.LastUpdated)
 	if err != nil {
 		if err == sql.ErrNoRows {
 			return persist.User{}, persist.ErrUserNotFound{Username: pUsername}
@@ -380,11 +420,11 @@ func (u *UserRepository) GetByUsername(pCtx context.Context, pUsername string) (
 
 }
 
-// GetByEmailIgnoringStatus gets the user with the given email address
-func (u *UserRepository) GetByEmailIgnoringStatus(pCtx context.Context, pEmail persist.Email) (persist.User, error) {
+// GetByEmail gets the user with the given email address
+func (u *UserRepository) GetByVerifiedEmail(pCtx context.Context, pEmail persist.Email) (persist.User, error) {
 
 	var user persist.User
-	err := u.getByEmailStmt.QueryRowContext(pCtx, pEmail.String()).Scan(&user.ID, &user.Deleted, &user.Version, &user.Username, &user.UsernameIdempotent, pq.Array(&user.Wallets), &user.Bio, &user.Traits, &user.Universal, &user.PrimaryWalletID, &user.CreationTime, &user.LastUpdated, &user.EmailVerified)
+	err := u.getByVerifiedEmailStmt.QueryRowContext(pCtx, pEmail.String()).Scan(&user.ID, &user.Deleted, &user.Version, &user.Username, &user.UsernameIdempotent, pq.Array(&user.Wallets), &user.Bio, &user.Traits, &user.Universal, &user.PrimaryWalletID, &user.CreationTime, &user.LastUpdated)
 	if err != nil {
 		if err == sql.ErrNoRows {
 			return persist.User{}, persist.ErrUserNotFound{Email: pEmail}
